Damage falloff and better tracing

+++ src/PlayerAbilities.cpp
@@ -5,13 +5,13 @@
 #include "ASGEGameLib/GameObjects/Player/HUD/PlayerAbilities.hpp"
 PlayerAbilities::PlayerAbilities(ASGE::Renderer* renderer, Player& _player, float x, float y) :
   player(_player), position(ASGE::Point2D(x, y)),
-  ability_icons(std::array<Sprite, 3>{Sprite(renderer, "data/images/ui/invisible_icon.png", 0),
-                                      Sprite(renderer, "data/images/ui/invisible_icon.png", 0),
-                                      Sprite(renderer, "data/images/ui/invisible_icon.png", 0)})
+  ability_icons(std::array<Sprite, 3>{ Sprite(renderer, "data/images/ui/invisible_icon.png", 0),
+                                       Sprite(renderer, "data/images/ui/invisible_icon.png", 0),
+                                       Sprite(renderer, "data/images/ui/invisible_icon.png", 0) })
 {
-  for(int i = 0; i < 3; i++)
+  for (size_t i = 0; i < 3; i++)
   {
-    ability_icons[i].dimensions(ASGE::Point2D(30,30));
+    ability_icons[i].dimensions(ASGE::Point2D(30, 30));
     ability_icons[i].position(ASGE::Point2D(position.x + (static_cast<float>(30 * i)), position.y));
   }
 }
@@ -20,7 +20,7 @@ void PlayerAbilities::update(InputTracker& input, float dt)
 {
   GameObject().update(input, dt);
 
-  if(player.is_invis)
+  if (player.is_invis)
   {
     ability_icons[0].colour(ASGE::COLOURS::GREEN);
   }
@@ -28,14 +28,14 @@ void PlayerAbilities::update(InputTracker& input, float dt)
   {
     ability_icons[0].colour(ASGE::COLOURS::WHITE);
   }
-  ability_icons[0].opacity(1 -player.invis_cooldown / 15);
+  ability_icons[0].opacity(1 - player.invis_cooldown / 15);
 }
 
 void PlayerAbilities::render(ASGE::Renderer* renderer)
 {
-  if(!player.is_dead)
+  if (!player.is_dead)
   {
-    for(auto& ability_icon : ability_icons)
+    for (auto& ability_icon : ability_icons)
     {
       ability_icon.render(renderer);
     }
